// src/app/models/mae-rubro.ts
export class MaeRubro {
  codEmpresa?: string;
  codRubro?: string;

  desRubro?: string;
  codTipoAuxi?: string;
  indProv?: string;
  indOp?: string;
  indCanc?: string;
  indRend?: string;
  indObra?: string;
  indNb?: string;
  indIn?: string;
  indFondo?: string;
  indCtaCte?: string;
  codDocumento?: string;
  indAplica?: string;
  indFacturacion?: string;
  indRent?: string;
}