// src/app/models/mae-tipo-gasto.ts
export class MaeTipoGasto {
  codEmpresa?: string;
  codRubro?: string;
  codTipoGasto?: string;

  codFlujo?: string;
  codTipoPlanilla?: string;
  numVerPlanCuentas?: string;
  codCuentaDolares?: string;
  codCuentaSoles?: string;
  desTipoGasto?: string;
  indEmpObr?: string;
  indCtaCte?: string;
  tipDebeHaber?: string;
  indFacturacion?: string;
  indDetraccion?: string;
  prcDet?: number;
  tipDetraccion?: string;
  tipFlujo?: string;
  codRubroAnt?: string;
  codTipoGastoAnt?: string;
  codRubroNew?: string;
  codTipoGastoNew?: string;
  codSunat?: string;
  codBienServicio?: string;
  indBolsa?: string;
  codCuentaDolaresAnt?: string;
  codCuentaSolesAnt?: string;
  tipMinimoDetrac?: string;
}